server/requirements.txt
Flask==2.1.0
Flask-CORS==3.0.10
pymongo==4.1.1
python-dotenv==0.19.2
Werkzeug==2.0.3
bson==0.5.10